Small Typo Fixes to Lessons 1 and 2
Here is the email from Brad I responded to…

As I mentioned to you, I found a few typos when I began reading the 221
book.  Here’s what I found in lessons 1 and 2 so far:

1.      Lesson 1.  In the “Help Desk” portion of the “Course Resources”
section, the phone number should be changed from 9000 to 1411.  I’m not
sure if the toll-free number in the book is correct or not.

2.      Lesson 2.  In the “Designed Experiments” section, Question 2
about the Salk video needs ellipses at the end to clarify that it is a
“complete the statement” question.  It should read “At that time,
parents were afraid to let their children…”  or maybe “….let their
children do what?”

3.      Lesson 2.  Question 10 about the Salk video says “Which of the
following actions is most likely at the conclusion of the Salk trial?”
And then there are no following actions listed.