Lesson 5 z-score typo fix...
A student contacted me and wrote:

We define an unusual observation to be something that happens less than
5% of the time. For normally distributed data, we determine if an
observation is unusual based on its z-score. We call an observation
unusual if z<2 or if z>2. In other words, we will call an event unusual
if the absolute value of its z-score is greater than 2.

I believe that it should read:

We call an observation unusual if z<-2 or if z>2