Small typo fixes to Lesson 9
Here is the email from Brad Garner:

“In the Body Temperatures part of the Lesson 9 chapter in the 221 book,
the P-value for the test is listed as 2.584 x 10^-11.  Using the
applet, the value I get is 2.5931x10^-11.

Also, at the end of the Ethan Allen explanation, in the next-to-last
paragraph, there is written 1.7217x10^-22, but the “times” symbol is
the “x” symbol for the variable x.  The times symbol should not be
italicized like that.”