Craig's suggestion to improve wording of a statement
Craig’s email:

There is a confusing statement in Lesson 10 of the textbook. Here is
the text:

"When we incorporate the margin of error, we estimate that the true
proportion of people who favor the Republican candidate is 49% ± 3%, or
in other words between 46% and 52%. For the Democratic candidate, we
get 48% to 54%. There is a lot of overlap in these numbers. In this
case, the political race is too close to know who might win."

I recommend simpler wording:

"When we consider the margin of error, we estimate that the true
proportion of people who favor the Republican candidate is 49% ± 3%, or
in other words between 46% and 52%. It is not obvious whether the
Republican candidate is favored by more or less than 50% of the voters.
In this case, the political race is too close to know who might win."

The original wording makes it sound like there are two confidence
intervals here, which is not the case.