Removed hyp test stuff, clarified sampling and data types
Removed all references to null and alternative hypotheses, but kept the
ideas behind both in the writing. Also removed references to p-value.
Improved examples and definitions, as well as formatting, of the five
sampling methods. Cleaned up the quantitative vs. categorical data
section to provide a little more clarification and instruction of what
a quantitative measurement is compared to a categorical label.