Update toolbox
Garrett,

In the toolbox, when the residuals are calculated in the Regression spreadsheet, Columns X and Y are formatted to three decimal places. This means the residual plot formats to three decimal places. For many data sets, this is not helpful. Will you please remove the formatting from these two columns?

While you are looking at the Toolbox, I have one more request. In the Chi-square sheet, the cells in Rows 5 & 6 are combined where you enter the data. It would be great if we could make it like the Expected Counts table, where the cells are not merged. The issue is that if you copy and paste data in the observed counts table, you cannot bring the column headers. It says you cannot do that because the copy area and paste area are the same size. If you need a demo or further clarification, please let me know.

Best,

Craig