Craig's updates
A few items:

Is it accidental that there are two versions of the toolbox on the main page of the textbook?

In the Linear Regression tab of the Toolbox, the degrees of freedom is on the left of the test statistic. Can we swap those? In every other tab that involves degrees of freedom, the test statistic is on the left and the degrees of freedom are on the right.

The degrees of freedom on the Linear Regression tab is set to 3 decimal places of precision. (This is a wee bit too precise for an integer value.) Could we please remove that formatting?

Since I am asking for a bunch of stuff, I will throw this in. This is just a matter of personal preference. On the one and two proportions page, I would love it if the cells for x and n were swapped. Why? We divide x by n when computing a proportion, so it is natural to think of x coming first and n coming second. This may be more trouble than it is worth.