Fix One Proportion Z-score
I checked the toolbox and book links in your other pages, the R textbook data links and on the main site, statistics.byuimath and they are all fine. From what I can tell it looks like the x and n were switch in the sandbox page toolbox, and the formula for zscore was not changed to match leading to division by x rather than n when calculating z score.