Add Probability to the online Lesson 1
After reviewing both Lesson 1’s in the online textbook, I think there are a couple of things that need to change in the online Lesson 1:

Remove the Class Meeting Section – a short two line section
Include the section on Probability that is in the campus Lesson 1 page.
Include the three rules of probability in the lesson summary at the bottom

Garrett, if you could make those changes, particularly with including the Probability pieces in the online Lesson 1, that would be great!

Sincerely,

Ryan